refactor: move parseJiraDescription and parseJiraSubtasks functions to jiraUtils.ts
